Fixed cluster after network split.

+++ raft-node/src/main/java/org/distributed/model/appendentries/AppendEntriesResponse.java
@@ -4,7 +4,7 @@ package org.distributed.model.appendentries;
  * @author Oleksandr Havrylenko
  **/
 public record AppendEntriesResponse(
-        int term,
+        long term,
         boolean success
 ) {
 }
